Improve markdown parsing
Resolves #12 and includes some refactoring for #7.
